+++ Demos/04-Teams/Demo-05/ConversationalBotTeamsToolkit/ConvBotWithTab/.env
@@ -1,0 +1,2 @@
+BotId=c303b380-6407-40b8-be3c-683442feeb2a
+BotPassword=ApfbNyBLdVba2-R4-48KV_0.2.3w6zFUB_
