+++ Demos/04-Teams/Demo-05/ConversationalBotTeamsToolkit/ConvBotWithTab/appPackage/outline.png
