--- Demos/04-Teams/Demo-05/ConversationalBotTeamsToolkit/ConversationalBot/.env
@@ -1,2 +1,0 @@
-BotId=d947ac5d-41f0-4a13-a244-2c48b1dc94e1
-BotPassword=rWa88L86lu5a_JK3l_63Cvyelq~EWC-Bu_
